python_service2/requirements.txt
black==22.3.0
isort==5.10.1
mypy==0.950
ipdb
opentelemetry-api==1.11.1
opentelemetry-exporter-otlp==1.11.1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-sdk==1.11.1
